requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Django==4.2.6
dj-database-url==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-phone-field==1.8.1
whitenoise==6.6.0